requirements.txt

fastapi
uvicorn
python-multipart
PyPDF2
openai
pytesseract
pdf2image
Pillow
beautifulsoup4
requests
uvicorn[standard]
python-dotenv
pydantic
boto3   
